Migrate raw_table to object_runtime (#1547)
