+++ frameworks/moveos-stdlib/src/natives/moveos_stdlib/mod.rs
@@ -7,7 +7,6 @@ pub mod event;
 pub mod json;
 pub mod move_module;
 pub mod object;
-pub mod raw_table;
 pub mod rlp;
 pub mod signer;
 pub mod test_helper;
