+++ moveos/moveos-object-runtime/src/lib.rs
@@ -2,3 +2,4 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod resolved_arg;
+pub mod runtime;
